Added testing for improved event and payload handling
